API_RAML
In the above example we first define the resourceType “collection,” and 
then call it into our resource using the type property. We are also taking 
advantage of three placeholders <<resourcePathName>> that are 
automatically filled with the resource name (“my-resource,” “resource-two”), 
#defined in our resources. Now, instead of having to write the entire resource 
#ach and every time, we can utilize this template, saving substantial
amounts of code and time.
Both “my-resource” and “resource-two” will now have a description and a 
GET method with 200, 301 and 400 responses. The 200 response returns 
back an application/json response with the example response we provided 
using the <<exampleGetResponse>> placeholder, and a redirect in the  
case of a 301 with <<exampleGetRedirect>>